MyProject/requirements.txt
Django==3.2
Django==4.2.5
djangorestframework==3.14.0
Requests==2.31.0
